requirements.test.txt
-r requirements.txt
pytest==8.3.1
pytest-cov==5.0.0
python-dotenv==1.0.1
pyyaml==6.0.1
